refactor: rename package structure from 'com.tech.springbaseproject' to 'com.tech.easy'
