+++ src/main/java/com/tech/easy/SpringBaseProjectApplication.java
@@ -1,0 +1,13 @@
+package com.tech.easy;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBaseProjectApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBaseProjectApplication.class, args);
+	}
+
+}
