+++ src/main/java/com/tech/easy/config/SwaggerConfig.java
@@ -1,0 +1,27 @@
+package com.tech.easy.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springdoc.core.models.GroupedOpenApi;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("Base Project API")
+                        .version("1.0")
+                        .description("API documentation for the base project"));
+    }
+
+    @Bean
+    public GroupedOpenApi publicApiV1() {
+        return GroupedOpenApi.builder()
+                .group("v1")
+                .pathsToMatch("/v1/**")
+                .build();
+    }
+}
