+++ src/main/java/com/tech/easy/domain/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.tech.easy.domain.repository;
+
+import com.tech.easy.domain.model.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    UserEntity findByUsername(String username);
+}
