+++ src/main/java/com/tech/easy/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.tech.easy.exception;
+
+public class UserNotFoundException extends BaseException {
+    public UserNotFoundException(Long userId) {
+        super(404, "User not found, id=" + userId);
+    }
+}
