--- src/main/java/com/tech/springbaseproject/SpringBaseProjectApplication.java
@@ -1,13 +1,0 @@
-package com.tech.springbaseproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringBaseProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringBaseProjectApplication.class, args);
-	}
-
-}
