--- src/main/java/com/tech/springbaseproject/common/Version.java
@@ -1,18 +1,0 @@
-package com.tech.springbaseproject.common;
-
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.web.bind.annotation.Mapping;
-
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Mapping
-public @interface Version {
-    int value(); // e.g., @Version(1) => /v1
-}
-
