--- src/main/java/com/tech/springbaseproject/config/CorrelationIdFilter.java
@@ -1,37 +1,0 @@
-package com.tech.springbaseproject.config;
-
-import jakarta.servlet.*;
-import jakarta.servlet.http.HttpServletRequest;
-import org.slf4j.MDC;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-
-import java.io.IOException;
-import java.util.UUID;
-
-@Component
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class CorrelationIdFilter implements Filter {
-    private static final String CORRELATION_ID_HEADER = "X-Correlation-ID";
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-
-        HttpServletRequest httpReq = (HttpServletRequest) request;
-        String correlationId = httpReq.getHeader(CORRELATION_ID_HEADER);
-
-        if (!StringUtils.hasText(correlationId)) {
-            correlationId = UUID.randomUUID().toString();
-        }
-
-        MDC.put("correlationId", correlationId);
-        try {
-            chain.doFilter(request, response);
-        } finally {
-            MDC.remove("correlationId");
-        }
-    }
-}
