--- src/main/java/com/tech/springbaseproject/config/InternationalizationConfig.java
@@ -1,25 +1,0 @@
-package com.tech.springbaseproject.config;
-
-import java.util.ArrayList;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import org.springframework.web.servlet.LocaleResolver;
-
-@Configuration
-public class InternationalizationConfig {
-
-    @Bean
-    public LocaleResolver localeResolver() {
-        return new CustomLocaleResolver();
-    }
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename("classpath:i18n/messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-}
