--- src/main/java/com/tech/springbaseproject/domain/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.tech.springbaseproject.domain.repository;
-
-import com.tech.springbaseproject.domain.model.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    UserEntity findByUsername(String username);
-}
