--- src/main/java/com/tech/springbaseproject/web/dto/UserResponse.java
@@ -1,20 +1,0 @@
-package com.tech.springbaseproject.web.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponse {
-    private Long id;
-    private String username;
-    private String email;
-
-    public UserResponse() {}
-    public UserResponse(Long id, String username, String email) {
-        this.id = id;
-        this.username = username;
-        this.email = email;
-    }
-    // getters/setters
-}
