+++ src/test/java/com/tech/easy/SpringBaseProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tech.easy;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBaseProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
