--- src/test/java/com/tech/springbaseproject/SpringBaseProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tech.springbaseproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringBaseProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
